+++ src/components/base/index.ts
@@ -3,4 +3,3 @@ export { default as Tag } from "./Tag";
 export { default as Text } from "./Text";
 export { default as Paragraph } from "./Paragraph";
 export { default as Img } from "./Img";
-export { default as IconTag } from "./IconTag";